pyproject.toml
[tool.poetry]
name = "discord-games-bot"
version = "0.1.0"
description = ""
authors = ["MateuszKubiszewski <matkub19@gmail.com>"]
readme = "README.md"
packages = [{include = "discord_games_bot"}]

[tool.poetry.dependencies]
python = "^3.10"
discord-py = "^2.1.0"
python-dotenv = "^0.21.0"
boto3 = "^1.26.40"
jsonpickle = "^3.0.1"
beautifulsoup4 = "^4.11.1"
pillow = "^9.3.0"
unidecode = "^1.3.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
